requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
causal-conv1d==1.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
grpcio==1.73.1
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.8
lightning-utilities==0.14.3
mamba-ssm==1.0.1
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.6.3
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
numpy==1.26.4
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
propcache==0.3.2
protobuf==4.23.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
pystrum==0.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.11.4
six==1.17.0
sympy==1.14.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
timm==0.4.12
tokenizers==0.21.2
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchmetrics==1.7.4
torchvision==0.16.1+cu121
tqdm==4.67.1
transformers==4.53.3
triton==2.1.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
